% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_carbon.R
\name{compute_carbon}
\alias{compute_carbon}
\title{compute_carbon}
\usage{
compute_carbon(currentbiomass, species, species.parm, springt, springp)
}
\arguments{
\item{currentbiomass}{(mgC)}

\item{species}{(name of species)}

\item{species.parm}{(data frame with species, maxrate (%C/yr), topt (C), pmax (mm), pmin(mm)}

\item{springt}{(C) springtime temperature}

\item{springp}{(mm) springtime rainfall}
}
\value{
growth (mgC/year)
}
\description{
computes growth given species, and spring temperature and precipitation
}

