% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ecobenefit.R
\name{compute_ecobenefit}
\alias{compute_ecobenefit}
\title{Forest Ecosystem Benefit Computer}
\usage{
compute_ecobenefit(tree, carbonprice, biodiversityprice, coeff.species.growth,
  clim, discount)
}
\arguments{
\item{tree}{dataframe with species and current biomass}

\item{carbonprice}{($) price paid for carbon}

\item{biodiversityprice}{($) price paid for biodiversity in a given year}

\item{clim}{dataframe with tmax, tmin and precip for each day}

\item{discount}{discount rates}

\item{paramters}{for growth model}
}
\value{
annual.benefit and NPV of all benefits over all years
}
\description{
compute_ecobenefit()
}
\details{
Computes an estimate of forest ecosystem benefits that include both biodiversity and carbon
}
\examples{



}

