% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_NPV.R
\name{compute_NPV}
\alias{compute_NPV}
\title{compute_NPV}
\usage{
compute_NPV(value, time, discount = 0.01)
}
\arguments{
\item{time}{in the future that cost/value occurs (years)}

\item{discount}{rate, default 0.01}

\item{value/cost}{($)}
}
\value{
value in $
}
\description{
compute net present value
}

