% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_simpson_index.R
\name{compute_simpson_index}
\alias{compute_simpson_index}
\title{Simpson's Species Diversity Index}
\usage{
compute_simpson_index(species)
}
\arguments{
\item{species}{list of species (names, or code)}
}
\value{
value of Species Diversity Index
}
\description{
Compute a species diversity index
}
\references{
http://www.tiem.utk.edu/~gross/bioed/bealsmodules/simpsonDI.html
}

