\name{esm237examples-package}
\alias{esm237examples-package}
\alias{esm237examples}
\docType{package}
\title{
\packageTitle{esm237examples}
}
\description{
\packageDescription{esm237examples}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{esm237examples}
\packageIndices{esm237examples}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{esm237examples}

Maintainer: \packageMaintainer{esm237examples}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
