% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spring.summary.R
\name{spring.summary}
\alias{spring.summary}
\title{Summary information about spring climate}
\usage{
spring.summary(clim.data, spring.months = c(4:6))
}
\arguments{
\item{clim.data}{data frame with columns tmax, tmin (C)
rain (precip in mm), year, month (integer), day}

\item{months}{(as integer) to include in spring; default 4,5,6}
}
\value{
returns a list containing, mean spring temperature (mean.springT, (C))
year with lowest spring temperature (coldest.spring (year))
mean spring precipitation (mean.springP (mm))
spring (as year) with highest precip (wettest.spring (year))
}
\description{
computes summary information about spring temperature and precipitation
}

